removes conversions

+++ flow/connectors/utils/partition/partition.go
@@ -54,9 +54,9 @@ func compareValues(prevEnd interface{}, start interface{}) int {
 			}
 		}
 	case uint32: //xmin
-		if prevEnd.(uint32) < uint32(v) {
+		if prevEnd.(uint32) < v {
 			return -1
-		} else if prevEnd.(uint32) > uint32(v) {
+		} else if prevEnd.(uint32) > v {
 			return 1
 		} else {
 			return 0
